# perf: lazy load workbox-build to reduce config loading time