I was thinking that vite's starting up time was slow in [my project](https://github.com/traPtitech/traQ_S-UI/).
I digged down a bit and found that evaluating `workbox-build` was making it slow.

This PR changes `workbox-build` to be lazy loaded.
Since `workbox-build` is not always used, I think it is ok to do this.

Without this PR, my project takes about 1300 - 1500ms to start up.
With this PR, it takes about 500 - 600ms.

These values are obtained from output with `DEBUG="vite:config" npm run dev` like `bundled config file loaded in 565.03ms +0ms`.