@userquin 2022-04-03
@sapphi-red 

sorry for the delay, testing with vitesse template also the dev startup from 2s to 1.25s, loading `workbox-build` on demand:

```shell
  vite v2.9.0 dev server running at:

  > Local: http://localhost:3333/
  > Network: use `--host` to expose

  ready in 1201ms.
```

with static import:
```shell
  vite v2.9.0 dev server running at:

  > Local: http://localhost:3333/
  > Network: use `--host` to expose

  ready in 2069ms.
```